requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
Pillow
picamera2